MUS-1880: Add Auth for `sendTimeStamp` Webhook
* JIRA: [MUS-1880](https://muselive.atlassian.net/browse/MUS-1880)
* Issue: #깃헙_이슈_번호

* **PR을 생성한 사람이** 머지를 진행합니다.
* 2개 이상의 approve가 필요합니다.

1. API GW와의 shared secret 기반 `sendTimeStamp` 업데이트 웹훅을 위한 HMAC authentication 추가
2. 디렉토리 정리, env 파일 추가

- [x] 정상 동작 확인
- [x] 코드 수정에 따른 주석, 문서 수정

<details><summary>리뷰어 도움말</summary>
<p>

> 리뷰 코멘트시 다음의 말머리를 사용해 주세요.

- [의견] - 더 나은 것으로 생각되는 코드가 있는 경우 제안
- [질문] - 설명이 부족하거나 이 부분만으로는 이유를 알 수 없는 부분에 설명을 부탁
- [확인] - 동작에는 지장이 없거나 적지만,수정이 필요하다고 생각되는 부분에 대한 확인 요청
- [이슈] - 문제가 발생할 것으로 예상됨
</p>
</details>
