+++ pull_request_template.md
@@ -1,0 +1,29 @@
+* JIRA: [티켓_번호](https://muselive.atlassian.net/browse/[티켓_번호])
+* Issue: #깃헙_이슈_번호
+
+## 머지 규칙
+* **PR을 생성한 사람이** 머지를 진행합니다.
+* 2개 이상의 approve가 필요합니다.
+
+## 작업 내용
+1. 작업 내용 요약 작성
+   > 왜 이렇게 해결했는지, 이 방법이 최선인 이유
+
+## 체크리스트
+- [ ] 정상 동작 확인
+- [ ] 코드 수정에 따른 주석, 문서 수정
+
+## 기타
+<details><summary>리뷰어 도움말</summary>
+<p>
+
+> 리뷰 코멘트시 다음의 말머리를 사용해 주세요.
+
+#### 머지 OK
+- [의견] - 더 나은 것으로 생각되는 코드가 있는 경우 제안
+- [질문] - 설명이 부족하거나 이 부분만으로는 이유를 알 수 없는 부분에 설명을 부탁
+#### 머지 전 확인/반영 부탁
+- [확인] - 동작에는 지장이 없거나 적지만,수정이 필요하다고 생각되는 부분에 대한 확인 요청
+- [이슈] - 문제가 발생할 것으로 예상됨
+</p>
+</details>
