+++ utils/constants/constants.go
@@ -3,8 +3,5 @@ package constants
 // Maximum number of device regstration tokens for single FCM multicast
 const FCM_TOKENS_THRESHOLD = 500
 
-// Webhook endpoint to trigger on notification sends
-const WEBHOOK_URL = "https://dev.clubmuse.live/api/webhooks"
-
 // Categories of hall-type shows
 const NOTI_CATEGORY_HALL = "hall"
